Generate services for all listeners
With https://github.com/ServiceWeaver/weaver/pull/388, we are able to
get for each component the list of corresponding listeners.

In this PR, we create a k8s service for each listener. By default, each
listener is private and we create a ClusterIP service. However, if the
names of the listeners are specified in the `[kube]` section of the
config, those listeners will be considered public and we will generate a
LoadBalancer service instead.