Add grafana integration (#40)
Grafana is a tool to visualize metrics, traces, and logs. It has a nice
UI and integration with N different data sources, including prometheus,
and jaeger.

Grafana should be the default way for someone to get insights into a
Service Weaver app when deployed with kube.

This PR launches a Grafana service and provides a Prometheus and a
Jaeger data source by default. We can simply open Grafana and see the
metrics, traces as exported by prometheus and jaeger.

In a future PR we will add integration with Loki to visualize the logs
as well.