Integrate with existing observability systems (#44)
* Integrate with existing observability systems

Before this PR, the Kube deployer will automatically start services to
provide observability to the user. We start Prometheus, Jaeger, Loki,
Promtail and Grafana.

While this is great for a simple deployment, in many production
scenarios, the user doesn't want to start some of these services (or
disable everything for testing/benchmarking performance). The user can
run its own custom services that scale well and have complicated
configs.

This PR enables the user to plugin custom observability services.
However, note that we are opinionated in terms of which services they
can use, namely Prometheus for metrics, Jaeger for traces, Loki/Promtail
for logs and Grafana for nice visualization. We don't enable integration
with other services yet.

* Addressed Srdjan's comments