Add cpu/mem reqs/limits as knobs.
Based on the latest Go survey (and interactions with many customers),
we've realized that enabling developers to control the resource
requirements for the running pods is very important.

This PR enables to control resource requests and resource limits for the
kube deployer.

This PR also enhances the `weaver kube deploy help` message to reflect
most of the knobs people can control right now when using the kube
deployer.