Removed internal_port from ReplicaSetConfig. (#77)
Previously, the `ReplicaSetConfig` proto contained an `internal_port`
field that configured the port on which a weavelet listened. This field
was not needed, as the weavelet port is a constant. There was no way for
a user to change the port, nor any reason for them to do so. This PR
removes the unneeded field.