Don't use kube.proto in kube.go. (#81)
Previously, `ReplicaSetConfig` in `kube.proto` was used in two ways. It
was serialized and passed to the babysitter, and it was used internally
by `kube.go`. This PR refactors `kube.go` to use a `deployment` struct
instead of a `ReplicaSetConfig`. This will enable me to simplify
`ReplicaSetConfig` in a future PR.