Add pod anti affinity rule by default
Right now, the Kubernetes scheduler can deploy replicas of the same
service on the same node. This is not desirable for HA. Also, it is
impossible for the user to specify any kind of affinity rules for the
pods.

This PR adds the ability to configure an affinity spec in the config.
Also, it adds by default an anti-affinity rule to make sure that the
Kubernetes scheduler is doing its best to schedule different replicas
for a service to different nodes.