Add app name to pods/services names (#113)
Right now, the user has to filter based on labels to see which pods are
from which app.

This PR adds the app name to the names of the pods and services.
While the names will become slightly longer, it makes it easier for
users to see which resources are from each app w/o filtering based on
labels. Also, this is consistent with what we do in the GKE deployer, where
the app name is also part of the resource names.